API: Capturing all transaction done by user

+++ FinanceApp.Server/FinanceApp.Infrastructure/Models/Accounts/AccountMetadataMedia.cs
@@ -8,15 +8,15 @@ namespace FinanceApp.Infrastructure.Models.Accounts
 {
     public class AccountMetadataMedia
     {
-        public double totalBalance { get; set; } = 0.0;
-        public double totalExpance { get; set; } = 0.0;
-        public double totalIncome { get; set; } = 0.0;
+        public decimal totalBalance { get; set; } = 0;
+        public decimal totalExpance { get; set; } = 0;
+        public decimal totalIncome { get; set; } = 0;
         public List<AccountMetadata> accounts { get; set; } = new List<AccountMetadata>();
     }
 
     public class AccountMetadata
     {
         public string name { get; set; }
-        public double totalBalance { get; set; } = 0.0;
+        public decimal totalBalance { get; set; } = 0;
     }
 }
