+++ FinanceApp.Server/FinanceApp.Infrastructure/Models/Transactions/TransactionsResponseMedia.cs
@@ -10,7 +10,7 @@ namespace FinanceApp.Infrastructure.Models.Transactions
     {
         public Guid accountId { get; set; }
         public Guid categoryId { get; set; }
-        public int amount { get; set; }
+        public decimal amount { get; set; }
         public string note { get; set; }
         public string transactionType { get; set; }
     }
